refactor(host): extract test modules from app/mod.rs and process_app/mod.rs (#1261)
Phase 1 of #1261: pure code motion, zero behavioral change.

- app/mod.rs: moved `mod tests` (582 lines) → app/tests.rs
- process_app/mod.rs: moved 11 test modules (1570 lines) → process_app/tests/

app/mod.rs: 4684→4103 lines (-12%)
process_app/mod.rs: 3544→1974 lines (-44%)
