+++ src/process_app/mod.rs
@@ -1890,403 +1890,8 @@ fn load_app_state(type_id: &str, workspace_root: &std::path::Path) -> serde_json
     serde_json::Value::Object(serde_json::Map::new())
 }
 
-// ── Tests ────────────────────────────────────────────────────────────────────
-
 #[cfg(test)]
-mod clipboard_tests {
-    //! Behavioural tests for the v3.2 clipboard / paste plumbing (#200, #146).
-    //!
-    //!   1. `egui::Event::Paste(text)` translates into `PlexiEvent::Paste`
-    //!      on the outbound queue (paste_event_forwarded_as_plexi_event).
-    //!   2. `DrawCommand::CopyToClipboard { text }` reaches egui's output
-    //!      command queue as `OutputCommand::CopyText` so the platform
-    //!      backend writes to the OS clipboard
-    //!      (copy_to_clipboard_drawcommand_calls_egui_copy).
-    //!
-    //! These exercise the host-side translation logic. End-to-end clipboard
-    //! integration with NSPasteboard / X11 / Wayland is verified via the
-    //! human-verification checklist in the PR — egui's backend is opaque
-    //! from a unit-test standpoint.
-    use super::*;
-    use crate::app_protocol::PlexiEvent;
-    use std::collections::HashSet;
-    use std::path::PathBuf;
-
-    /// Build a minimal `ProcessApp` for tests. Mirrors the helper in
-    /// `text_input_tests` — spawns `/bin/sh -c "sleep 1"` so lifecycle
-    /// machinery is happy, then ignores the subprocess.
-    fn make_app() -> Option<ProcessApp> {
-        let sh = ["/bin/sh", "/usr/bin/sh"]
-            .iter()
-            .find(|p| std::path::Path::new(p).exists())
-            .map(PathBuf::from)?;
-        let workspace_root = std::env::temp_dir();
-        ProcessApp::launch(
-            "test_clipboard",
-            "Test Clipboard",
-            &sh,
-            &workspace_root,
-            &["-c".to_string(), "sleep 1".to_string()],
-            workspace_root.clone(),
-            HashSet::new(),
-            false,
-            None,
-        )
-        .ok()
-    }
-
-    #[test]
-    fn paste_event_forwarded_as_plexi_event() {
-        // Drive a synthesised `egui::Event::Paste("hello")` through the
-        // pane's `handle_key`. The expected outcome is one
-        // `PlexiEvent::Paste { text: "hello" }` on the outbound event
-        // queue. No `Key`/`Text` events should be synthesised.
-        let Some(mut app) = make_app() else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-
-        let mut input = egui::InputState::default();
-        input.events.push(egui::Event::Paste("hello".to_string()));
-
-        let consumed = app.handle_key(&input);
-        assert!(consumed, "handle_key must consume Paste events");
-
-        let paste_events: Vec<_> = app
-            .outbound_events
-            .iter()
-            .filter(|e| matches!(e, PlexiEvent::Paste { .. }))
-            .collect();
-        assert_eq!(
-            paste_events.len(),
-            1,
-            "expected exactly one Paste event, got {paste_events:?}"
-        );
-        match paste_events[0] {
-            PlexiEvent::Paste { text } => assert_eq!(text, "hello"),
-            other => panic!("expected Paste, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn copy_to_clipboard_drawcommand_calls_egui_copy() {
-        // Verify the wired path: `ControlCommand::CopyToClipboard { text }` →
-        // `egui::Context::copy_text(text)` → `OutputCommand::CopyText` on
-        // the platform output. We construct a fresh egui Context, mirror
-        // the one-line dispatch from `ProcessApp::handle_control_command()`,
-        // and inspect the platform output for the CopyText command. If the
-        // dispatch ever changes shape (e.g. a different egui method), this
-        // test forces the breakage to surface.
-        use crate::app_protocol::ControlCommand;
-        let ctx = egui::Context::default();
-        let cmd = ControlCommand::CopyToClipboard {
-            text: "selected snippet".to_string(),
-        };
-
-        // This mirrors the exact branch in `ProcessApp::handle_control_command()`.
-        // Keep it in sync — if you refactor the dispatch, refactor this too.
-        match cmd {
-            ControlCommand::CopyToClipboard { text } => ctx.copy_text(text),
-            _ => panic!("test setup error"),
-        }
-
-        // Drain platform output and look for CopyText.
-        let mut found = None;
-        ctx.output_mut(|o| {
-            for cmd in &o.commands {
-                if let egui::OutputCommand::CopyText(text) = cmd {
-                    found = Some(text.clone());
-                }
-            }
-        });
-        assert_eq!(
-            found.as_deref(),
-            Some("selected snippet"),
-            "CopyToClipboard must emit OutputCommand::CopyText with the right text"
-        );
-    }
-
-    #[test]
-    fn crash_overlay_c_key_copies_report() {
-        use crate::testing::HostHarness;
-        use crate::pane::{AppRuntime, Pane};
-
-        let mut h = HostHarness::new();
-        let pane = h.add_test_pane();
-
-        // Force lifecycle to Crashed and inject known stderr lines.
-        {
-            let win = &mut h.app.windows[0];
-            let Some(Pane::App(app_pane)) = win.panes.get_mut(&pane) else {
-                panic!("expected App pane");
-            };
-            let AppRuntime::Process(proc) = &mut app_pane.runtime else {
-                panic!("expected Process runtime");
-            };
-            proc.lifecycle.on_process_exited(); // → Crashed
-            let mut buf = proc.recent_stderr.lock().unwrap();
-            buf.push_back("Traceback (most recent call last):".to_string());
-            buf.push_back("  File \"app.py\", line 42, in run".to_string());
-            buf.push_back("ZeroDivisionError: division by zero".to_string());
-        }
-
-        // One frame to trigger the overlay and stamp crashed_at.
-        h.run_frames(1);
-
-        // Send C — no modifier.
-        h.key(egui::Key::C, egui::Modifiers::NONE);
-
-        // Check the clipboard from the last frame's platform output.
-        let copy_cmd = h.last_platform_output.commands.iter().find_map(|cmd| {
-            if let egui::OutputCommand::CopyText(text) = cmd {
-                Some(text.clone())
-            } else {
-                None
-            }
-        });
-        let report = copy_cmd.expect("pressing C on crash overlay must write to clipboard");
-
-        assert!(
-            report.contains("=== Plexi Crash Report ==="),
-            "report must have header: {report}"
-        );
-        assert!(
-            report.contains("crashed"),
-            "report must name the state: {report}"
-        );
-        assert!(
-            report.contains("ZeroDivisionError"),
-            "report must contain stderr lines: {report}"
-        );
-        assert!(
-            report.contains("Traceback"),
-            "report must contain all stderr lines: {report}"
-        );
-    }
-}
-
-#[cfg(test)]
-mod text_input_tests {
-    //! Buffer-state tests for the v3.1 host-owned TextInput primitive
-    //! (issue #283). Covered behaviours:
-    //!   1. Buffer persists across frames until submit.
-    //!   2. Enter emits `PlexiEvent::TextSubmitted` with the buffered value.
-    //!   3. Submit clears the buffer (so the field is empty on the next emit).
-    //!   4. Pane resize (which triggers re-render but not buffer touch) does
-    //!      not wipe the buffer.
-    //!
-    //! These exercise the persistent-state contract — the egui rendering
-    //! layer is verified end-to-end by the human-verification checklist.
-    //! Keeping the unit tests pure makes them deterministic and fast.
-    use super::*;
-    use crate::app_protocol::PlexiEvent;
-    use std::collections::HashSet;
-    use std::path::PathBuf;
-    use std::process::Command;
-
-    /// Build a `ProcessApp` for tests that doesn't touch real I/O. We
-    /// spawn `/bin/sh -c true` — the cheapest valid subprocess — and
-    /// then ignore the lifecycle / draw threads. The app's stdin/stdout
-    /// are real but we never write to them in these tests.
-    fn make_app() -> Option<ProcessApp> {
-        let sh = ["/bin/sh", "/usr/bin/sh"]
-            .iter()
-            .find(|p| std::path::Path::new(p).exists())
-            .map(PathBuf::from)?;
-        let workspace_root = std::env::temp_dir();
-        // -c true exits immediately. The lifecycle reader threads will
-        // observe stdout EOF and flip Crashed, which is fine — these
-        // tests don't read lifecycle state.
-        let _ = Command::new(&sh); // sanity — silences unused-import warnings on some configs
-        ProcessApp::launch(
-            "test_text_input",
-            "Test Text Input",
-            &sh,
-            &workspace_root,
-            &["-c".to_string(), "sleep 1".to_string()],
-            workspace_root.clone(),
-            HashSet::new(),
-            false,
-            None,
-        )
-        .ok()
-    }
-
-    #[test]
-    fn text_input_buffer_persists_across_frames() {
-        let Some(mut app) = make_app() else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-        // Simulate two frames where the user has typed "hel" then "hello"
-        // by directly manipulating the buffer the way egui's TextEdit
-        // would across two ui() ticks.
-        app.render_session.text_input_buffers
-            .insert("note".to_string(), "hel".to_string());
-        // ... another frame happens (no submit) ...
-        // Buffer must still be there.
-        assert_eq!(
-            app.render_session.text_input_buffers.get("note").map(String::as_str),
-            Some("hel"),
-            "buffer should survive between frames"
-        );
-        app.render_session.text_input_buffers
-            .insert("note".to_string(), "hello".to_string());
-        assert_eq!(
-            app.render_session.text_input_buffers.get("note").map(String::as_str),
-            Some("hello")
-        );
-    }
-
-    #[test]
-    fn enter_emits_text_submitted_with_buffered_value() {
-        let Some(mut app) = make_app() else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-        app.render_session.text_input_buffers
-            .insert("note".to_string(), "hello world".to_string());
-
-        app.submit_text_input("note");
-
-        let evt = app.outbound_events.pop_back().expect("event queued");
-        match evt {
-            PlexiEvent::TextSubmitted { id, value } => {
-                assert_eq!(id, "note");
-                assert_eq!(value, "hello world");
-            }
-            other => panic!("expected TextSubmitted, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn submit_clears_buffer() {
-        let Some(mut app) = make_app() else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-        app.render_session.text_input_buffers
-            .insert("note".to_string(), "draft".to_string());
-        app.submit_text_input("note");
-        assert!(
-            !app.render_session.text_input_buffers.contains_key("note"),
-            "buffer must be cleared after submit (default UX)"
-        );
-    }
-
-    #[test]
-    fn submit_on_empty_buffer_emits_empty_value() {
-        let Some(mut app) = make_app() else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-        // No prior buffer — Enter on a fresh TextInput is a valid case
-        // (e.g. user immediately presses Enter without typing).
-        app.submit_text_input("note");
-        let evt = app.outbound_events.pop_back().expect("event queued");
-        match evt {
-            PlexiEvent::TextSubmitted { id, value } => {
-                assert_eq!(id, "note");
-                assert_eq!(value, "");
-            }
-            other => panic!("expected TextSubmitted, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn text_input_buffer_survives_pane_resize() {
-        let Some(mut app) = make_app() else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-        // Pane resize is a `last_size` change in `ui()`. The buffer is
-        // owned by `render_session.text_input_buffers` and never touched by resize
-        // logic. Simulate the resize bookkeeping and assert the buffer
-        // is untouched.
-        app.render_session.text_input_buffers
-            .insert("note".to_string(), "midway".to_string());
-        // Resize bookkeeping (mirrors what `ui()` does on size delta).
-        app.last_size = egui::vec2(800.0, 600.0);
-        // No buffer mutation should happen here — just last_size changes.
-        assert_eq!(
-            app.render_session.text_input_buffers.get("note").map(String::as_str),
-            Some("midway"),
-            "resize must not wipe the host-owned text buffer"
-        );
-    }
-
-    #[test]
-    fn distinct_ids_keep_independent_buffers() {
-        let Some(mut app) = make_app() else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-        app.render_session.text_input_buffers
-            .insert("a".to_string(), "alpha".to_string());
-        app.render_session.text_input_buffers
-            .insert("b".to_string(), "beta".to_string());
-        app.submit_text_input("a");
-        assert_eq!(
-            app.render_session.text_input_buffers.get("b").map(String::as_str),
-            Some("beta"),
-            "submitting one input must not affect another id"
-        );
-    }
-}
-
-#[cfg(test)]
-mod render_session_tests {
-    use super::*;
-    use std::collections::HashSet;
-    use std::path::PathBuf;
-
-    fn make_app() -> Option<ProcessApp> {
-        let sh = ["/bin/sh", "/usr/bin/sh"]
-            .iter()
-            .find(|p| std::path::Path::new(p).exists())
-            .map(PathBuf::from)?;
-        let workspace_root = std::env::temp_dir();
-        ProcessApp::launch(
-            "test_render_session",
-            "Test RenderSession",
-            &sh,
-            &workspace_root,
-            &["-c".to_string(), "sleep 1".to_string()],
-            workspace_root.clone(),
-            HashSet::new(),
-            false,
-            None,
-        ).ok()
-    }
-
-    #[test]
-    fn render_session_submit_produces_event() {
-        let Some(mut app) = make_app() else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-        app.render_session.text_input_buffers.insert("x".to_string(), "hello".to_string());
-        app.submit_text_input("x");
-        let evt = app.outbound_events.pop_back().expect("event queued");
-        match evt {
-            crate::app_protocol::PlexiEvent::TextSubmitted { id, value } => {
-                assert_eq!(id, "x");
-                assert_eq!(value, "hello");
-            }
-            other => panic!("expected TextSubmitted, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn render_session_process_app_has_no_text_input_fields() {
-        // Compile-time proof: ProcessApp::render_session owns the state.
-        // This test just exercises the field path — if mod.rs still had
-        // text_input_buffers directly on ProcessApp this wouldn't compile.
-        let Some(mut app) = make_app() else { return; };
-        app.render_session.text_input_buffers.insert("k".to_string(), "v".to_string());
-        assert!(app.render_session.text_input_buffers.contains_key("k"));
-    }
-}
+mod tests;
 
 impl Drop for ProcessApp {
     fn drop(&mut self) {
@@ -2388,1232 +1993,5 @@ impl Drop for ProcessApp {
                 let _ = child.wait();
             }
         }
-    }
-}
-
-#[cfg(test)]
-mod ai_tests {
-    //! Routing tests for the v3.3 `ai.query` broker capability (#284).
-    //!
-    //! Two paths under test:
-    //!   1. App without `ai.query` capability — synchronous denial response
-    //!      lands on `outbound_events`. No broker dispatch occurs.
-    //!   2. App with `ai.query` capability — broker is invoked once and
-    //!      its response surfaces on `http_rx` as `PlexiEvent::AiResponse`.
-    //!
-    //! The mock broker (`CannedBroker`) records every call so the granted
-    //! path also confirms that the routing layer forwarded the right
-    //! `model_tier`, `system`, and `messages` payload to the broker.
-    use super::*;
-    use crate::app_protocol::{AiMessage, HostCommand, ModelTier, PlexiEvent};
-    use crate::plexi_ai::broker::{AiBroker, AiBrokerRequest, AiBrokerResponse};
-    use std::collections::HashSet;
-    use std::path::PathBuf;
-    use std::sync::{Arc, Mutex};
-
-    /// Test broker: records every dispatch and returns a canned response.
-    struct CannedBroker {
-        seen: Arc<Mutex<Vec<AiBrokerRequest>>>,
-        response: AiBrokerResponse,
-    }
-
-    impl AiBroker for CannedBroker {
-        fn dispatch(&self, request: AiBrokerRequest) -> AiBrokerResponse {
-            self.seen.lock().unwrap().push(request);
-            self.response.clone()
-        }
-    }
-
-    fn make_app(capabilities: HashSet<Capability>) -> Option<ProcessApp> {
-        let sh = ["/bin/sh", "/usr/bin/sh"]
-            .iter()
-            .find(|p| std::path::Path::new(p).exists())
-            .map(PathBuf::from)?;
-        let workspace_root = std::env::temp_dir();
-        ProcessApp::launch(
-            "test_ai",
-            "Test AI",
-            &sh,
-            &workspace_root,
-            &["-c".to_string(), "sleep 1".to_string()],
-            workspace_root.clone(),
-            capabilities,
-            false,
-            None,
-        )
-        .ok()
-    }
-
-    #[test]
-    fn denied_app_gets_capability_denied_response() {
-        // App without `ai.query` capability: route_command must immediately
-        // queue an AiResponse with the canonical "capability denied" error
-        // — synchronously, without ever invoking the broker.
-        let Some(mut app) = make_app(HashSet::new()) else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-
-        // Inject a broker that would *panic* if called. The denied path
-        // must short-circuit before reaching dispatch.
-        struct PanicBroker;
-        impl AiBroker for PanicBroker {
-            fn dispatch(&self, _: AiBrokerRequest) -> AiBrokerResponse {
-                panic!("denied path must never call the broker");
-            }
-        }
-        app.ai_broker = Arc::new(PanicBroker);
-
-        app.route_command(HostCommand::AiQuery {
-            request_id: "req-denied".to_string(),
-            model_tier: ModelTier::Low,
-            system: "system".to_string(),
-            messages: vec![AiMessage {
-                role: "user".to_string(),
-                content: "hi".to_string(),
-            }],
-            tools: vec![],
-        });
-
-        // Denied path is synchronous — the response is on outbound_events
-        // immediately, no thread, no http_rx wait.
-        let resp = app
-            .outbound_events
-            .iter()
-            .find(|e| matches!(e, PlexiEvent::AiResponse { .. }))
-            .expect("expected AiResponse on outbound queue");
-        match resp {
-            PlexiEvent::AiResponse {
-                request_id,
-                content,
-                tokens_in,
-                tokens_out,
-                error,
-            } => {
-                assert_eq!(request_id, "req-denied");
-                assert!(content.is_none());
-                assert_eq!(*tokens_in, 0);
-                assert_eq!(*tokens_out, 0);
-                let err = error.as_ref().expect("error must be set on denial");
-                assert!(
-                    err.contains("capability denied"),
-                    "denial message must say `capability denied`: {err}"
-                );
-                assert!(
-                    err.contains("ai.query"),
-                    "denial message must name the capability: {err}"
-                );
-            }
-            other => panic!("expected AiResponse, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn granted_app_dispatches_to_broker() {
-        // App WITH `ai.query` granted: route_command must spawn a worker
-        // that calls broker.dispatch exactly once with the right payload,
-        // and the broker's response must arrive as a PlexiEvent::AiResponse
-        // on http_rx.
-        let mut caps = HashSet::new();
-        caps.insert(Capability::AiQuery);
-        let Some(mut app) = make_app(caps) else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-
-        let seen: Arc<Mutex<Vec<AiBrokerRequest>>> = Arc::new(Mutex::new(Vec::new()));
-        app.ai_broker = Arc::new(CannedBroker {
-            seen: Arc::clone(&seen),
-            response: AiBrokerResponse::ok("Pong.".to_string(), 12, 4),
-        });
-
-        app.route_command(HostCommand::AiQuery {
-            request_id: "req-ok".to_string(),
-            model_tier: ModelTier::High,
-            system: "be terse".to_string(),
-            messages: vec![AiMessage {
-                role: "user".to_string(),
-                content: "ping".to_string(),
-            }],
-            tools: vec![],
-        });
-
-        // Worker thread is spawned — wait briefly for response to arrive
-        // on http_rx. 2s is generous; canned broker is in-memory so the
-        // typical wait is microseconds.
-        let event = app
-            .http_rx
-            .recv_timeout(std::time::Duration::from_secs(2))
-            .expect("broker response must arrive on http_rx within 2s");
-
-        match event {
-            PlexiEvent::AiResponse {
-                request_id,
-                content,
-                tokens_in,
-                tokens_out,
-                error,
-            } => {
-                assert_eq!(request_id, "req-ok");
-                assert_eq!(content.as_deref(), Some("Pong."));
-                assert_eq!(tokens_in, 12);
-                assert_eq!(tokens_out, 4);
-                assert!(error.is_none());
-            }
-            other => panic!("expected AiResponse, got {other:?}"),
-        }
-
-        // Broker must have been invoked exactly once with the correct payload.
-        let calls = seen.lock().unwrap();
-        assert_eq!(calls.len(), 1, "broker must be called exactly once");
-        assert_eq!(calls[0].app_id, "test_ai");
-        assert_eq!(calls[0].model_tier, ModelTier::High);
-        assert_eq!(calls[0].system, "be terse");
-        assert_eq!(calls[0].messages.len(), 1);
-        assert_eq!(calls[0].messages[0].content, "ping");
-    }
-}
-
-#[cfg(test)]
-mod midi_tests {
-    //! Routing tests for the v3.4 CoreMIDI capability (#320).
-    //!
-    //! Two paths under test:
-    //!   1. App without `midi.in` / `midi.out` — synchronous denial response
-    //!      lands on `outbound_events`. No device dispatch occurs.
-    //!   2. App with the capability — `MockMidiDevice` records the open and
-    //!      the routing layer queues `MidiInputOpened` on success.
-    use super::*;
-    use crate::app_protocol::{HostCommand, PlexiEvent};
-    use crate::midi::MockMidiDevice;
-    use std::collections::HashSet;
-    use std::path::PathBuf;
-    use std::sync::Arc;
-
-    fn make_app(capabilities: HashSet<Capability>) -> Option<ProcessApp> {
-        let sh = ["/bin/sh", "/usr/bin/sh"]
-            .iter()
-            .find(|p| std::path::Path::new(p).exists())
-            .map(PathBuf::from)?;
-        let workspace_root = std::env::temp_dir();
-        ProcessApp::launch(
-            "test_midi",
-            "Test MIDI",
-            &sh,
-            &workspace_root,
-            &["-c".to_string(), "sleep 1".to_string()],
-            workspace_root.clone(),
-            capabilities,
-            false,
-            None,
-        )
-        .ok()
-    }
-
-    #[test]
-    fn denied_app_gets_capability_denied_response() {
-        // App without `midi.in`: route_command must immediately queue
-        // a MidiInputError with "capability denied" and never touch the
-        // device. The send path mirrors the same contract for `midi.out`.
-        let Some(mut app) = make_app(HashSet::new()) else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-
-        let mock = Arc::new(MockMidiDevice::new());
-        app.midi_device = Arc::clone(&mock) as Arc<dyn crate::midi::MidiDevice>;
-
-        app.route_command(HostCommand::OpenMidiInput {
-            port_id: "mock-input-1".to_owned(),
-            pipe_id: "midi-in-pipe".to_owned(),
-        });
-
-        let evt = app
-            .outbound_events
-            .iter()
-            .find(|e| matches!(e, PlexiEvent::MidiInputError { .. }))
-            .expect("expected MidiInputError on outbound queue");
-        match evt {
-            PlexiEvent::MidiInputError { pipe_id, error } => {
-                assert_eq!(pipe_id, "midi-in-pipe");
-                assert!(
-                    error.contains("capability denied"),
-                    "denial must say `capability denied`: {error}"
-                );
-                assert!(
-                    error.contains("midi.in"),
-                    "denial must name the capability: {error}"
-                );
-            }
-            other => panic!("expected MidiInputError, got {other:?}"),
-        }
-
-        // The mock must NOT have an active session — the denied path
-        // short-circuits before open_input is called.
-        assert!(
-            mock.injected_sinks
-                .lock()
-                .expect("mock midi sinks poisoned")
-                .is_empty(),
-            "denied path must not open the MIDI input"
-        );
-
-        // SendMidi without `midi.out` is the same shape.
-        app.route_command(HostCommand::SendMidi {
-            port_id: "mock-output-1".to_owned(),
-            bytes: vec![0x90, 0x3C, 0x64],
-        });
-        let evt = app
-            .outbound_events
-            .iter()
-            .find(|e| matches!(e, PlexiEvent::MidiSendError { .. }))
-            .expect("expected MidiSendError on outbound queue");
-        match evt {
-            PlexiEvent::MidiSendError { port_id, error } => {
-                assert_eq!(port_id, "mock-output-1");
-                assert!(
-                    error.contains("capability denied"),
-                    "denial must say `capability denied`: {error}"
-                );
-                assert!(
-                    error.contains("midi.out"),
-                    "denial must name the capability: {error}"
-                );
-            }
-            other => panic!("expected MidiSendError, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn granted_app_dispatches_open_input_to_device() {
-        // App WITH `midi.in` granted: route_command must open the input on
-        // the device, register a sink, and queue MidiInputOpened on
-        // outbound_events. With `midi.out` granted, SendMidi must dispatch
-        // the bytes to the mock's `sent` log.
-        let mut caps = HashSet::new();
-        caps.insert(Capability::MidiIn);
-        caps.insert(Capability::MidiOut);
-        let Some(mut app) = make_app(caps) else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-
-        let mock = Arc::new(MockMidiDevice::new());
-        app.midi_device = Arc::clone(&mock) as Arc<dyn crate::midi::MidiDevice>;
-
-        app.route_command(HostCommand::OpenMidiInput {
-            port_id: "mock-input-1".to_owned(),
-            pipe_id: "midi-in-pipe".to_owned(),
-        });
-
-        // PipeOpened arrives BEFORE MidiInputOpened so the app can connect
-        // the unix socket before the first byte arrives.
-        let pipe_opened = app
-            .outbound_events
-            .iter()
-            .position(|e| matches!(e, PlexiEvent::PipeOpened { .. }))
-            .expect("expected PipeOpened");
-        let midi_opened = app
-            .outbound_events
-            .iter()
-            .position(|e| matches!(e, PlexiEvent::MidiInputOpened { .. }))
-            .expect("expected MidiInputOpened");
-        assert!(
-            pipe_opened < midi_opened,
-            "PipeOpened must precede MidiInputOpened so the app's socket connection races first"
-        );
-
-        // The device must have a registered sink for the port.
-        assert!(
-            mock.injected_sinks
-                .lock()
-                .expect("mock midi sinks poisoned")
-                .contains_key("mock-input-1"),
-            "open_input must have registered a sink"
-        );
-
-        // SendMidi path: dispatches one note-on to the mock output log.
-        app.route_command(HostCommand::SendMidi {
-            port_id: "mock-output-1".to_owned(),
-            bytes: vec![0x90, 0x3C, 0x64],
-        });
-        let log = mock.sent.lock().expect("mock sent poisoned").clone();
-        let entries = log
-            .get("mock-output-1")
-            .expect("mock-output-1 entries must exist after SendMidi");
-        assert_eq!(entries.len(), 1, "exactly one send dispatched");
-        assert_eq!(entries[0], vec![0x90u8, 0x3C, 0x64]);
-    }
-}
-
-#[cfg(test)]
-mod video_tests {
-    //! Routing tests for the v3.4 video substrate (#345).
-    //!
-    //! Two paths under test:
-    //!   1. App without `video.playback` — synchronous denial response on
-    //!      `outbound_events`. No device dispatch.
-    //!   2. App with the capability — `MockVideoDecoder` opens the source,
-    //!      the routing layer queues `VideoOpenAck` and pumps frames.
-    use super::*;
-    use crate::app_protocol::{HostCommand, PlexiEvent};
-    use crate::video::{MockVideoDecoder, MockVideoDecoderConfig};
-    use std::collections::HashSet;
-    use std::path::PathBuf;
-    use std::sync::Arc;
-
-    fn make_app(capabilities: HashSet<Capability>) -> Option<ProcessApp> {
-        let sh = ["/bin/sh", "/usr/bin/sh"]
-            .iter()
-            .find(|p| std::path::Path::new(p).exists())
-            .map(PathBuf::from)?;
-        let workspace_root = std::env::temp_dir();
-        ProcessApp::launch(
-            "test_video",
-            "Test Video",
-            &sh,
-            &workspace_root,
-            &["-c".to_string(), "sleep 1".to_string()],
-            workspace_root.clone(),
-            capabilities,
-            false,
-            None,
-        )
-        .ok()
-    }
-
-    #[test]
-    fn denied_app_gets_capability_denied_response() {
-        // App without `video.playback`: route_command must immediately queue
-        // a VideoOpenError with "capability denied" and never touch the
-        // decoder.
-        let Some(mut app) = make_app(HashSet::new()) else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-
-        let mock = Arc::new(MockVideoDecoder::new(MockVideoDecoderConfig::default()));
-        app.video_device = Arc::clone(&mock) as Arc<dyn crate::video::VideoDecoder>;
-
-        app.route_command(HostCommand::OpenVideo {
-            request_id: "req-denied".to_owned(),
-            source: "mock://gradient".to_owned(),
-            pipe_id: "video-stream".to_owned(),
-        });
-
-        let evt = app
-            .outbound_events
-            .iter()
-            .find(|e| matches!(e, PlexiEvent::VideoOpenError { .. }))
-            .expect("expected VideoOpenError on outbound queue");
-        match evt {
-            PlexiEvent::VideoOpenError { request_id, error } => {
-                assert_eq!(request_id, "req-denied");
-                assert!(
-                    error.contains("capability denied"),
-                    "denial must say `capability denied`: {error}"
-                );
-                assert!(
-                    error.contains("video.playback"),
-                    "denial must name the capability: {error}"
-                );
-            }
-            other => panic!("expected VideoOpenError, got {other:?}"),
-        }
-
-        // The denial path must not produce a VideoOpenAck.
-        assert!(
-            !app
-                .outbound_events
-                .iter()
-                .any(|e| matches!(e, PlexiEvent::VideoOpenAck { .. })),
-            "denied path must not produce a VideoOpenAck"
-        );
-        assert!(
-            app.video_handles.is_empty(),
-            "denied path must not register a handle"
-        );
-    }
-
-    #[test]
-    fn granted_app_dispatches_open_to_decoder() {
-        // App WITH `video.playback`: route_command must open the decoder,
-        // queue PipeOpened then VideoOpenAck, and register a handle. Then
-        // SetVideoState dispatches into the handle without panicking, and
-        // CloseVideo tears it down cleanly.
-        let mut caps = HashSet::new();
-        caps.insert(Capability::VideoPlayback);
-        let Some(mut app) = make_app(caps) else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-
-        let mock = Arc::new(MockVideoDecoder::new(MockVideoDecoderConfig {
-            width: 16,
-            height: 8,
-            fps: 30.0,
-            duration_ms: 5_000,
-        }));
-        app.video_device = Arc::clone(&mock) as Arc<dyn crate::video::VideoDecoder>;
-
-        app.route_command(HostCommand::OpenVideo {
-            request_id: "req-1".to_owned(),
-            source: "mock://gradient".to_owned(),
-            pipe_id: "video-stream".to_owned(),
-        });
-
-        // PipeOpened arrives BEFORE VideoOpenAck so the app can connect the
-        // unix socket before the first frame.
-        let pipe_opened = app
-            .outbound_events
-            .iter()
-            .position(|e| matches!(e, PlexiEvent::PipeOpened { .. }))
-            .expect("expected PipeOpened");
-        let video_ack = app
-            .outbound_events
-            .iter()
-            .position(|e| matches!(e, PlexiEvent::VideoOpenAck { .. }))
-            .expect("expected VideoOpenAck");
-        assert!(
-            pipe_opened < video_ack,
-            "PipeOpened must precede VideoOpenAck so the app's socket connection races first"
-        );
-
-        // Pull the ack out and confirm the dimensions match the mock config.
-        let ack_handle_id = match &app.outbound_events[video_ack] {
-            PlexiEvent::VideoOpenAck {
-                handle_id,
-                width,
-                height,
-                fps,
-                duration_ms,
-                request_id,
-            } => {
-                assert_eq!(request_id, "req-1");
-                assert_eq!(*width, 16);
-                assert_eq!(*height, 8);
-                assert!((*fps - 30.0).abs() < 0.01);
-                assert_eq!(*duration_ms, 5_000);
-                *handle_id
-            }
-            _ => unreachable!(),
-        };
-        assert!(
-            app.video_handles.contains_key(&ack_handle_id),
-            "open must register a handle"
-        );
-
-        // SetVideoState — pause then play, neither should panic and no error
-        // event should arrive.
-        app.route_command(HostCommand::SetVideoState {
-            handle_id: ack_handle_id,
-            state: crate::video::VideoState::Pause,
-        });
-        app.route_command(HostCommand::SetVideoState {
-            handle_id: ack_handle_id,
-            state: crate::video::VideoState::Play,
-        });
-        app.route_command(HostCommand::SetVideoState {
-            handle_id: ack_handle_id,
-            state: crate::video::VideoState::Seek { position_ms: 1_000 },
-        });
-        // No additional VideoOpenError must have been queued.
-        let errors = app
-            .outbound_events
-            .iter()
-            .filter(|e| matches!(e, PlexiEvent::VideoOpenError { .. }))
-            .count();
-        assert_eq!(errors, 0, "set_state must not produce VideoOpenError");
-
-        // CloseVideo tears down the handle and unregisters the pipe id map.
-        app.route_command(HostCommand::CloseVideo {
-            handle_id: ack_handle_id,
-        });
-        assert!(
-            !app.video_handles.contains_key(&ack_handle_id),
-            "close must drop the handle"
-        );
-        assert!(
-            !app.video_pipe_ids.contains_key(&ack_handle_id),
-            "close must unregister the pipe id mapping"
-        );
-    }
-}
-
-// ── Canvas Terminal Binding Primitives — routing-layer tests (#78) ───────────
-//
-// These tests pin the `process_app::routing` behaviour for the five binding
-// primitives. The HOST-level dispatch (creating the terminal pane, injecting
-// bytes into the PTY) lives in `app/canvas_bindings.rs` and is exercised end-
-// to-end by the POC `canvas-terminal-bindings-test` app.
-//
-// Each test asserts one of two things:
-//   - Without `terminal.bindings`: the dispatch is a synchronous capability-
-//     deny path. No `AppCommand` is enqueued; for the request/response
-//     shapes (RequestLinkedTerminal, RequestCommandPreview) a sentinel
-//     event lands on `outbound_events` so the SDK's blocking helper
-//     unblocks instead of hanging.
-//   - With `terminal.bindings` granted: the matching `AppCommand` lands
-//     on `pending_commands` for the parent host to drain, populated with
-//     the right fields and `sender_pane_id` correctly stamped.
-#[cfg(test)]
-mod canvas_bindings_tests {
-    use super::*;
-    use crate::app_protocol::{
-        ArtifactOpenMode, HostCommand, PathTokenMode, PlexiEvent,
-    };
-    use std::collections::HashSet;
-    use std::path::PathBuf;
-
-    fn make_app(capabilities: HashSet<Capability>) -> Option<ProcessApp> {
-        let sh = ["/bin/sh", "/usr/bin/sh"]
-            .iter()
-            .find(|p| std::path::Path::new(p).exists())
-            .map(PathBuf::from)?;
-        let workspace_root = std::env::temp_dir();
-        let mut app = ProcessApp::launch(
-            "test_bindings",
-            "Test Bindings",
-            &sh,
-            &workspace_root,
-            &["-c".to_string(), "sleep 1".to_string()],
-            workspace_root.clone(),
-            capabilities,
-            false,
-            None,
-        )
-        .ok()?;
-        // Stamp a non-zero pane id so the AppCommand sender_pane_id is
-        // distinguishable from "unset".
-        app.set_pane_id(7);
-        Some(app)
-    }
-
-    // ── Capability denial paths ─────────────────────────────────────────
-
-    #[test]
-    fn denied_app_request_linked_terminal_emits_sentinel_event() {
-        let Some(mut app) = make_app(HashSet::new()) else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-        app.route_command(HostCommand::RequestLinkedTerminal {
-            request_id: "req-0".to_string(),
-            cwd: None,
-            label: None,
-        });
-        // Sentinel event so the SDK's blocking helper unblocks.
-        let event = app
-            .outbound_events
-            .iter()
-            .find(|e| matches!(e, PlexiEvent::LinkedTerminalReady { .. }))
-            .expect("denied path must emit LinkedTerminalReady sentinel");
-        match event {
-            PlexiEvent::LinkedTerminalReady {
-                request_id,
-                terminal_pane_id,
-            } => {
-                assert_eq!(request_id, "req-0");
-                assert_eq!(
-                    *terminal_pane_id, 0,
-                    "sentinel pane id (0) signals capability denied"
-                );
-            }
-            other => panic!("expected LinkedTerminalReady, got {other:?}"),
-        }
-        // Must NOT have queued an AppCommand for the host to act on.
-        assert!(
-            !app.pending_commands.iter().any(|c| matches!(
-                c,
-                AppCommand::RequestLinkedTerminal { .. }
-            )),
-            "denied path must not enqueue RequestLinkedTerminal"
-        );
-    }
-
-    #[test]
-    fn denied_app_run_in_linked_terminal_drops_silently() {
-        let Some(mut app) = make_app(HashSet::new()) else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-        app.route_command(HostCommand::RunInLinkedTerminal {
-            terminal_pane_id: 42,
-            command: "ls".to_string(),
-            echo: true,
-        });
-        // No event — fire-and-forget verb has no response shape.
-        // Must NOT enqueue the AppCommand.
-        assert!(
-            !app.pending_commands.iter().any(|c| matches!(
-                c,
-                AppCommand::RunInLinkedTerminal { .. }
-            )),
-            "denied path must drop RunInLinkedTerminal without dispatch"
-        );
-    }
-
-    #[test]
-    fn denied_app_request_command_preview_emits_empty_cwd_sentinel() {
-        let Some(mut app) = make_app(HashSet::new()) else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-        app.route_command(HostCommand::RequestCommandPreview {
-            request_id: "req-9".to_string(),
-            terminal_pane_id: 42,
-            command: "rm -rf .git".to_string(),
-        });
-        let event = app
-            .outbound_events
-            .iter()
-            .find(|e| matches!(e, PlexiEvent::CommandPreview { .. }))
-            .expect("denied path must emit CommandPreview sentinel");
-        match event {
-            PlexiEvent::CommandPreview {
-                request_id,
-                command,
-                would_run_in_cwd,
-            } => {
-                assert_eq!(request_id, "req-9");
-                assert_eq!(command, "rm -rf .git");
-                assert!(
-                    would_run_in_cwd.is_empty(),
-                    "denied path must return empty cwd: got {would_run_in_cwd:?}"
-                );
-            }
-            other => panic!("expected CommandPreview, got {other:?}"),
-        }
-    }
-
-    // ── Granted-path AppCommand enqueue ─────────────────────────────────
-
-    #[test]
-    fn granted_app_dispatches_request_linked_terminal() {
-        let mut caps = HashSet::new();
-        caps.insert(Capability::TerminalBindings);
-        let Some(mut app) = make_app(caps) else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-        app.route_command(HostCommand::RequestLinkedTerminal {
-            request_id: "req-ok".to_string(),
-            cwd: Some("/tmp/foo".to_string()),
-            label: Some("bindings demo".to_string()),
-        });
-        // No synchronous event — granted path defers to the host.
-        assert!(app
-            .outbound_events
-            .iter()
-            .find(|e| matches!(e, PlexiEvent::LinkedTerminalReady { .. }))
-            .is_none());
-        // The AppCommand lands on pending_commands with sender_pane_id
-        // stamped (set_pane_id(7)).
-        let cmd = app
-            .pending_commands
-            .iter()
-            .find_map(|c| {
-                if let AppCommand::RequestLinkedTerminal {
-                    sender_pane_id,
-                    request_id,
-                    cwd,
-                    label,
-                } = c
-                {
-                    Some((*sender_pane_id, request_id.clone(), cwd.clone(), label.clone()))
-                } else {
-                    None
-                }
-            })
-            .expect("granted path must enqueue AppCommand::RequestLinkedTerminal");
-        assert_eq!(cmd.0, 7, "sender_pane_id must come from app's pane_id");
-        assert_eq!(cmd.1, "req-ok");
-        assert_eq!(cmd.2.as_deref(), Some("/tmp/foo"));
-        assert_eq!(cmd.3.as_deref(), Some("bindings demo"));
-    }
-
-    #[test]
-    fn granted_app_dispatches_run_in_linked_terminal() {
-        let mut caps = HashSet::new();
-        caps.insert(Capability::TerminalBindings);
-        let Some(mut app) = make_app(caps) else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-        app.route_command(HostCommand::RunInLinkedTerminal {
-            terminal_pane_id: 42,
-            command: "ls -la".to_string(),
-            echo: true,
-        });
-        let cmd = app
-            .pending_commands
-            .iter()
-            .find_map(|c| {
-                if let AppCommand::RunInLinkedTerminal {
-                    terminal_pane_id,
-                    command,
-                    echo,
-                    ..
-                } = c
-                {
-                    Some((*terminal_pane_id, command.clone(), *echo))
-                } else {
-                    None
-                }
-            })
-            .expect("granted path must enqueue RunInLinkedTerminal");
-        assert_eq!(cmd.0, 42);
-        assert_eq!(cmd.1, "ls -la");
-        assert!(cmd.2);
-    }
-
-    #[test]
-    fn granted_app_dispatches_insert_path_token() {
-        let mut caps = HashSet::new();
-        caps.insert(Capability::TerminalBindings);
-        let Some(mut app) = make_app(caps) else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-        app.route_command(HostCommand::InsertPathToken {
-            terminal_pane_id: 42,
-            path: "/tmp/x".to_string(),
-            mode: PathTokenMode::Replace,
-        });
-        let mode = app
-            .pending_commands
-            .iter()
-            .find_map(|c| {
-                if let AppCommand::InsertPathToken { mode, .. } = c {
-                    Some(*mode)
-                } else {
-                    None
-                }
-            })
-            .expect("granted path must enqueue InsertPathToken");
-        assert_eq!(mode, PathTokenMode::Replace);
-    }
-
-    #[test]
-    fn granted_app_dispatches_open_artifact() {
-        let mut caps = HashSet::new();
-        caps.insert(Capability::TerminalBindings);
-        let Some(mut app) = make_app(caps) else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-        app.route_command(HostCommand::OpenArtifact {
-            path: "/tmp/x".to_string(),
-            mode: ArtifactOpenMode::RevealInFinder,
-        });
-        let mode = app
-            .pending_commands
-            .iter()
-            .find_map(|c| {
-                if let AppCommand::OpenArtifact { mode, .. } = c {
-                    Some(*mode)
-                } else {
-                    None
-                }
-            })
-            .expect("granted path must enqueue OpenArtifact");
-        assert_eq!(mode, ArtifactOpenMode::RevealInFinder);
-    }
-}
-
-#[cfg(test)]
-mod reload_tests {
-    //! Reload-path tests for hot reload (#83).
-    //!
-    //! Two paths under test:
-    //!   1. Drop on a well-behaved subprocess sends `Shutdown` and reaps
-    //!      cleanly within the 2s timeout (no SIGTERM escalation).
-    //!   2. Drop on an app that ignores Shutdown escalates to SIGTERM
-    //!      and still reaps within the 1s SIGTERM timeout.
-    //!
-    //! These exercise the `Drop for ProcessApp` machinery directly — the
-    //! reload glue in `pane_ops::create::reload_app_pane` relies on
-    //! `Drop` for the Shutdown→wait→kill sequence. Replacing the runtime
-    //! field naturally drops the old `ProcessApp`.
-    use super::*;
-    use std::collections::HashSet;
-    use std::path::PathBuf;
-
-    fn make_sh_app(args: &[&str]) -> Option<ProcessApp> {
-        let sh = ["/bin/sh", "/usr/bin/sh"]
-            .iter()
-            .find(|p| std::path::Path::new(p).exists())
-            .map(PathBuf::from)?;
-        let workspace_root = std::env::temp_dir();
-        let owned: Vec<String> = args.iter().map(|s| s.to_string()).collect();
-        ProcessApp::launch(
-            "test_reload",
-            "Test Reload",
-            &sh,
-            &workspace_root,
-            &owned,
-            workspace_root.clone(),
-            HashSet::new(),
-            false,
-            None,
-        )
-        .ok()
-    }
-
-    /// Reload contract: dropping a `ProcessApp` reaps the underlying
-    /// child. `Shutdown` is best-effort over a stdio that the child
-    /// likely isn't even reading; the timed escalation guarantees the
-    /// reap happens regardless.
-    #[test]
-    fn drop_reaps_well_behaved_subprocess_within_window() {
-        let Some(app) = make_sh_app(&["-c", "sleep 5"]) else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-        // The child PID must exist while `app` is alive.
-        let pid = app
-            .process
-            .as_ref()
-            .map(|c| c.id())
-            .expect("child should be running");
-        let start = std::time::Instant::now();
-        drop(app);
-        let elapsed = start.elapsed();
-
-        // Drop completed — under 4s ceiling (2s shutdown + 1s SIGTERM + slack).
-        assert!(
-            elapsed < std::time::Duration::from_secs(4),
-            "drop must complete within escalation window, took {elapsed:?}"
-        );
-
-        // Verify the OS released the PID (kill 0 to test process existence).
-        // On Unix, kill -0 returns -1 with ESRCH if the process doesn't exist.
-        #[cfg(unix)]
-        {
-            // Sleep a tick so the OS gets a chance to fully reap.
-            std::thread::sleep(std::time::Duration::from_millis(50));
-            let alive = unsafe { libc::kill(pid as libc::pid_t, 0) };
-            // kill returning -1 means the process is gone (ESRCH).
-            assert_eq!(
-                alive, -1,
-                "child pid {pid} must be reaped after drop, got {alive}"
-            );
-        }
-    }
-
-    /// Reload force-kill contract: a subprocess that ignores `Shutdown`
-    /// (no stdin reader, just `sleep`) must still be reaped via the
-    /// SIGTERM escalation.
-    #[test]
-    fn drop_force_kills_unresponsive_subprocess() {
-        // `sleep 30` doesn't read stdin and ignores SIGPIPE — the only
-        // way it dies is SIGTERM/SIGKILL.
-        let Some(app) = make_sh_app(&["-c", "exec sleep 30"]) else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-        let pid = app
-            .process
-            .as_ref()
-            .map(|c| c.id())
-            .expect("child should be running");
-        let start = std::time::Instant::now();
-        drop(app);
-        let elapsed = start.elapsed();
-
-        // The 2s shutdown window expires (sleep ignores it), then
-        // SIGTERM reaps within 1s. Total well under 4s.
-        assert!(
-            elapsed < std::time::Duration::from_secs(4),
-            "force-kill must complete within escalation window, took {elapsed:?}"
-        );
-        assert!(
-            elapsed >= std::time::Duration::from_millis(1900),
-            "shutdown wait should give the well-behaved app time to exit; elapsed={elapsed:?}"
-        );
-
-        #[cfg(unix)]
-        {
-            std::thread::sleep(std::time::Duration::from_millis(50));
-            let alive = unsafe { libc::kill(pid as libc::pid_t, 0) };
-            assert_eq!(
-                alive, -1,
-                "unresponsive child pid {pid} must be force-reaped after drop"
-            );
-        }
-    }
-
-    // ── stream child cleanup on drop (#675) ────────────────────────────────────
-
-    #[test]
-    fn drop_cancels_active_stream_children() {
-        let Some(mut app) = make_sh_app(&["-c", "sleep 5"]) else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-
-        // Spawn a long-running child to simulate a StreamProcess.
-        let mut stream_child = std::process::Command::new("/bin/sh")
-            .args(["-c", "exec sleep 60"])
-            .stdout(std::process::Stdio::piped())
-            .stderr(std::process::Stdio::piped())
-            .spawn()
-            .expect("spawn stream child");
-        let stream_pid = stream_child.id();
-
-        // Insert into stream_handles — mirrors what StreamProcess does.
-        app.stream_handles.insert(
-            "test-stream".to_string(),
-            StreamHandle {
-                cancel: Arc::new(std::sync::atomic::AtomicBool::new(false)),
-                pid: stream_pid,
-            },
-        );
-
-        drop(app);
-
-        // The SIGKILL escalation thread may have already reaped via waitpid,
-        // so wait() can return ECHILD — both outcomes prove the child is gone.
-        match stream_child.wait() {
-            Ok(status) => assert!(
-                !status.success(),
-                "stream child should have been killed by signal, got: {status}"
-            ),
-            Err(e) if e.raw_os_error() == Some(libc::ECHILD) => {
-                // Already reaped by the SIGKILL escalation thread — expected.
-            }
-            Err(e) => panic!("unexpected wait error: {e}"),
-        }
-
-        #[cfg(unix)]
-        {
-            let alive = unsafe { libc::kill(stream_pid as libc::pid_t, 0) };
-            assert_eq!(
-                alive, -1,
-                "stream child pid {stream_pid} must be gone after drop"
-            );
-        }
-    }
-
-    // ── render coalescing (issue #368) ────────────────────────────────────────
-
-    /// Verifies render-event coalescing: N back-to-back `PlexiEvent::Render`
-    /// calls must produce at most 1 `FlushRender` token in the channel (so
-    /// a burst never fills the queue and silently drops itself), and a
-    /// subsequent non-render event must still reach the subprocess.
-    ///
-    /// Strategy: rather than round-tripping through a subprocess, we inspect
-    /// the shared `render_slot` / `render_in_queue` Arcs directly after the
-    /// calls. This is race-free because `send_event` writes them synchronously
-    /// on the caller's thread before returning.
-    ///
-    /// Covered invariants:
-    ///   1. After N renders, `render_in_queue` is true (exactly one token queued).
-    ///   2. `render_slot` contains the *last* render's payload.
-    ///   3. A Key event sent after the renders does not clear or corrupt the slot.
-    ///   4. `event_tx` is still Some (no spurious disconnection).
-    #[test]
-    fn render_events_coalesced_non_render_events_preserved() {
-        let Some(mut app) = make_sh_app(&["-c", "sleep 5"]) else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-
-        let rect = crate::app_protocol::Rect { x: 0.0, y: 0.0, w: 800.0, h: 600.0 };
-
-        // Send 5 Render events back-to-back.
-        for frame_id in 1u64..=5 {
-            app.send_event(&PlexiEvent::Render { frame_id, rect: rect.clone() });
-        }
-
-        // After the burst, exactly one FlushRender token must be in the channel.
-        // `render_in_queue` is true while the token is queued / not yet drained.
-        assert!(
-            app.render_in_queue.load(Ordering::Relaxed),
-            "render_in_queue must be true after a burst of Render events"
-        );
-
-        // render_slot must hold the *latest* (frame_id=5) payload, not an earlier one.
-        {
-            let slot = app.render_slot.lock().unwrap();
-            let payload = slot.as_deref().expect("render_slot must be populated");
-            assert!(
-                payload.contains("\"frame_id\":5"),
-                "render_slot must contain the latest frame_id (5), got: {payload}"
-            );
-        }
-
-        // A Key event after the burst must be accepted without error.
-        app.send_event(&PlexiEvent::Key {
-            key: "j".to_string(),
-            modifiers: crate::app_protocol::Modifiers::default(),
-        });
-
-        // event_tx must still be live — the Key was enqueued successfully.
-        assert!(
-            app.event_tx.is_some(),
-            "event_tx must remain Some after sending a Key event"
-        );
-    }
-
-    /// Sanity check: `launch_process` is re-entrant for the same id —
-    /// hot reload calls it on every reload.
-    #[test]
-    fn launch_process_is_reentrant_for_same_app() {
-        let Some(a) = make_sh_app(&["-c", "sleep 0.1"]) else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-        let Some(b) = make_sh_app(&["-c", "sleep 0.1"]) else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-        let id_a = a.process.as_ref().map(|c| c.id());
-        let id_b = b.process.as_ref().map(|c| c.id());
-        assert!(id_a.is_some());
-        assert!(id_b.is_some());
-        assert_ne!(
-            id_a, id_b,
-            "two launches of the same app must produce distinct PIDs"
-        );
-    }
-}
-
-#[cfg(test)]
-mod app_state_tests {
-    use super::*;
-
-    #[test]
-    fn load_app_state_returns_empty_when_no_files() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let result = load_app_state("test-app", dir.path());
-        assert_eq!(result, serde_json::Value::Object(serde_json::Map::new()));
-    }
-
-    #[test]
-    fn load_app_state_reads_workspace_file_over_global() {
-        let ws_dir = tempfile::tempdir().expect("workspace tempdir");
-        let state_dir = ws_dir.path().join(".plexi").join("app_states");
-        std::fs::create_dir_all(&state_dir).expect("mkdir");
-        let state_path = state_dir.join("my-app.json");
-        std::fs::write(&state_path, r#"{"interval_idx":3}"#).expect("write");
-
-        let result = load_app_state("my-app", ws_dir.path());
-        assert_eq!(result["interval_idx"], serde_json::json!(3));
-    }
-
-    #[test]
-    fn load_app_state_migrates_old_app_state_dir() {
-        let ws_dir = tempfile::tempdir().expect("workspace tempdir");
-        let old_dir = ws_dir.path().join(".plexi").join("app_state");
-        std::fs::create_dir_all(&old_dir).expect("mkdir");
-        std::fs::write(old_dir.join("my-app.json"), r#"{"migrated":true}"#).expect("write");
-
-        let result = load_app_state("my-app", ws_dir.path());
-        assert_eq!(result["migrated"], serde_json::json!(true));
-        // Old dir must be gone, new dir must exist.
-        assert!(!old_dir.exists(), "old app_state dir should have been renamed");
-        assert!(ws_dir.path().join(".plexi").join("app_states").exists());
-    }
-}
-
-#[cfg(test)]
-mod env_isolation_tests {
-    //! Proves that user-global secrets (stored as `plexi:user:*`) are NOT
-    //! injected into app subprocess environments. The only path to a secret
-    //! is `HostCommand::SecretGet` through the brokered capability check.
-    //!
-    //! Strategy: add a canary directly to the Command (simulating what the old
-    //! unconditional list_user_secrets() injection did), then apply env_clear()
-    //! + whitelist (what ProcessApp::launch now does), and assert the canary
-    //! is absent from the subprocess. Rust's Command builder strips explicit
-    //! .env() additions that precede .env_clear(), so no host-process env
-    //! mutation is needed — no thread-safety concerns.
-
-    use std::process::Command;
-
-    const WHITELIST: &[&str] = &["HOME", "PATH", "LANG", "LC_ALL", "TERM", "USER", "SHELL"];
-    const CANARY_KEY: &str = "PLEXI_SECRET_CANARY";
-    const CANARY_VAL: &str = "secret_must_not_leak";
-
-    #[test]
-    fn user_global_secrets_not_injected_into_subprocess_env() {
-        let sh = ["/bin/sh", "/usr/bin/sh"]
-            .iter()
-            .find(|p| std::path::Path::new(p).exists())
-            .copied();
-        let Some(sh) = sh else {
-            eprintln!("skipping: no /bin/sh available");
-            return;
-        };
-
-        // Add the canary first (models what the old injection loop wrote),
-        // then env_clear() which strips it. env(k,v) before env_clear() is
-        // removed by env_clear() — verified empirically via Rust std behavior.
-        let output = Command::new(sh)
-            .arg("-c")
-            .arg(format!("echo \"${{{}:-ABSENT}}\"", CANARY_KEY))
-            .env(CANARY_KEY, CANARY_VAL)
-            .env_clear()
-            .envs(
-                WHITELIST
-                    .iter()
-                    .filter_map(|k| std::env::var(k).ok().map(|v| (*k, v))),
-            )
-            .output()
-            .expect("sh spawn failed");
-
-        let stdout = String::from_utf8_lossy(&output.stdout);
-        assert_eq!(
-            stdout.trim(),
-            "ABSENT",
-            "user-global secret must not appear in subprocess env; got: {stdout:?}"
-        );
-    }
-}
-
-#[cfg(test)]
-mod image_cache_tests {
-    use super::*;
-
-    #[test]
-    fn image_cache_loads_png() {
-        // Write a 1×1 red PNG to a temp dir using the `image` crate (avoids
-        // embedding raw bytes that could be subtly invalid).
-        let dir = tempfile::tempdir().expect("tempdir");
-        let png_path = dir.path().join("test.png");
-        let mut img = image::RgbaImage::new(1, 1);
-        img.put_pixel(0, 0, image::Rgba([255, 0, 0, 255]));
-        img.save(&png_path).expect("save png");
-
-        let mut cache = image_cache::ImageCache::new();
-        cache.request("test.png", dir.path());
-
-        // Give the background thread time to load.
-        std::thread::sleep(std::time::Duration::from_millis(200));
-        cache.poll(&egui::Context::default());
-
-        assert!(
-            cache.get("test.png").is_some(),
-            "expected image to be loaded"
-        );
-    }
-
-    #[test]
-    fn image_cache_missing_file_is_error() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let mut cache = image_cache::ImageCache::new();
-        cache.request("nonexistent.png", dir.path());
-
-        std::thread::sleep(std::time::Duration::from_millis(200));
-        cache.poll(&egui::Context::default());
-
-        assert!(
-            matches!(
-                cache.state("nonexistent.png"),
-                Some(image_cache::CachedImage::Error(_))
-            ),
-            "expected Error state for missing file"
-        );
     }
 }
