+++ src/process_app/tests/mod.rs
@@ -1,0 +1,11 @@
+mod clipboard_tests;
+mod text_input_tests;
+mod render_session_tests;
+mod ai_tests;
+mod midi_tests;
+mod video_tests;
+mod canvas_bindings_tests;
+mod reload_tests;
+mod app_state_tests;
+mod env_isolation_tests;
+mod image_cache_tests;
